Get rid of Google maps API key embedded in the HTML source
GCP sent me an email warning against this practice. This change conforms to best security practice. If I ever want to resurrect this project, I'll need to make sure to add a route in the backend that makes the google maps API call with an appropriate dev key, so that the key isn't published in the HTML source.